refactor(kb): split subcommands into separate files
Create src/cli/commands/kb/ directory with one file per subcommand:
- index.ts: thin router that registers all subcommands and re-exports
- toggle.ts: enable/disable/status + addKbHook/removeKbHook/hasKbHook
- list.ts: list subcommand handler
- check.ts: check subcommand handler
- create.ts: create subcommand handler
- refresh.ts: refresh subcommand handler
- remove.ts: remove subcommand handler
- shared.ts: featureKb module reference, exitOnInvalidSlug, getWorktreePath

kb.ts becomes a backward-compat re-export shim for existing callers
(init.ts, uninstall.ts, tests) that import from commands/kb.js.
