+++ src/cli/commands/kb.ts
@@ -1,547 +1,6 @@
-import { Command } from 'commander';
-import { promises as fs } from 'fs';
-import * as path from 'path';
-import * as p from '@clack/prompts';
-import color from 'picocolors';
-import { createRequire } from 'module';
-import { fileURLToPath } from 'url';
-import { isClaudeCliAvailable } from '../utils/cli.js';
-import { getGitRoot } from '../utils/git.js';
-import type { HookMatcher, Settings } from '../utils/hooks.js';
-import { getClaudeDirectory, getDevFlowDirectory } from '../utils/paths.js';
-import { readManifest, writeManifest } from '../utils/manifest.js';
-export type { SidecarData } from '../utils/sidecar.js';
-export { readSidecar } from '../utils/sidecar.js';
-import { runKbAgent } from '../utils/kb-agent.js';
-
-const __filename = fileURLToPath(import.meta.url);
-const __dirname = path.dirname(__filename);
-
-/** @internal */
-const _require = createRequire(import.meta.url);
-
-
-interface FeatureKbModule {
-  listKBs: (worktreePath: string) => Array<{ slug: string; name: string; directories: string[]; lastUpdated: string; referencedFiles?: string[]; description?: string; createdBy?: string }>;
-  checkAllStaleness: (worktreePath: string) => Record<string, { stale: boolean; changedFiles: string[] }>;
-  checkStaleness: (worktreePath: string, slug: string) => { stale: boolean; changedFiles: string[] };
-  findOverlapping: (worktreePath: string, changedFiles: string[]) => string[];
-  removeEntry: (worktreePath: string, slug: string) => void;
-  validateSlug: (slug: string) => void;
-  updateIndex: (worktreePath: string, entry: { slug: string; name: string; description?: string; directories: string[]; referencedFiles: string[]; createdBy?: string }, lockTimeoutMs?: number) => void;
-}
-
-// dist/cli/commands/kb.js → ../../.. → project root (where scripts/ lives)
-const featureKb: FeatureKbModule = _require(
-  path.join(__dirname, '..', '..', '..', 'scripts', 'hooks', 'lib', 'feature-kb.cjs')
-);
-
 /**
- * Validate a KB slug and exit with an error message if invalid.
- * Centralises the repeated try/catch pattern across create/refresh/remove.
+ * @deprecated Import from './kb/index.js' directly.
+ * This module is a compatibility shim preserved for callers that
+ * reference the old `commands/kb.js` path (e.g., tests, init.ts).
  */
-function exitOnInvalidSlug(slug: string): void {
-  try {
-    featureKb.validateSlug(slug);
-  } catch (err) {
-    p.log.error(err instanceof Error ? err.message : String(err));
-    process.exit(1);
-  }
-}
-
-/**
- * Get the git root for the current directory, or cwd if not in a git repo.
- */
-async function getWorktreePath(): Promise<string> {
-  return (await getGitRoot()) ?? process.cwd();
-}
-
-const KB_HOOK_MARKER = 'session-end-kb-refresh';
-
-/**
- * Add the KB SessionEnd hook to settings JSON.
- * Idempotent — returns unchanged JSON if hook already exists.
- */
-export function addKbHook(settingsJson: string, devflowDir: string): string {
-  if (hasKbHook(settingsJson)) {
-    return settingsJson;
-  }
-
-  const settings: Settings = JSON.parse(settingsJson);
-
-  if (!settings.hooks) {
-    settings.hooks = {};
-  }
-
-  const hookCommand = `${path.join(devflowDir, 'scripts', 'hooks', 'run-hook')} session-end-kb-refresh`;
-
-  const newEntry: HookMatcher = {
-    hooks: [
-      {
-        type: 'command',
-        command: hookCommand,
-        timeout: 10,
-      },
-    ],
-  };
-
-  if (!settings.hooks.SessionEnd) {
-    settings.hooks.SessionEnd = [];
-  }
-
-  settings.hooks.SessionEnd.push(newEntry);
-
-  return JSON.stringify(settings, null, 2) + '\n';
-}
-
-/**
- * Remove the KB hook from settings JSON.
- * Idempotent — returns unchanged JSON if hook not present.
- * Preserves other hooks. Cleans empty arrays/objects.
- */
-export function removeKbHook(settingsJson: string): string {
-  const settings: Settings = JSON.parse(settingsJson);
-  let changed = false;
-
-  const matchers = settings.hooks?.SessionEnd;
-  if (matchers) {
-    const filtered = matchers.filter(
-      (m) => !m.hooks.some((h) => h.command.includes(KB_HOOK_MARKER)),
-    );
-    if (filtered.length < matchers.length) changed = true;
-    if (filtered.length === 0) {
-      delete settings.hooks!.SessionEnd;
-    } else {
-      settings.hooks!.SessionEnd = filtered;
-    }
-  }
-
-  if (!changed) {
-    return settingsJson;
-  }
-
-  if (settings.hooks && Object.keys(settings.hooks).length === 0) {
-    delete settings.hooks;
-  }
-
-  return JSON.stringify(settings, null, 2) + '\n';
-}
-
-/**
- * Check if the KB hook is registered in settings JSON or parsed Settings object.
- */
-export function hasKbHook(input: string | Settings): boolean {
-  const settings: Settings = typeof input === 'string' ? JSON.parse(input) : input;
-  return settings.hooks?.SessionEnd?.some((matcher) =>
-    matcher.hooks.some((h) => h.command.includes(KB_HOOK_MARKER)),
-  ) ?? false;
-}
-
-export const kbCommand = new Command('kb')
-  .description('Manage per-feature knowledge bases')
-  .option('--enable', 'Enable per-feature knowledge bases')
-  .option('--disable', 'Disable per-feature knowledge bases')
-  .option('--status', 'Show KB feature status')
-  .action(async (options: { enable?: boolean; disable?: boolean; status?: boolean }) => {
-    if (!options.enable && !options.disable && !options.status) return;
-
-    const worktreePath = await getWorktreePath();
-    const claudeDir = getClaudeDirectory();
-    const devflowDir = getDevFlowDirectory();
-    const settingsPath = path.join(claudeDir, 'settings.json');
-
-    if (options.enable) {
-      p.intro(color.cyan('Enable Feature Knowledge Bases'));
-
-      // Create .features/index.json if missing
-      const featuresDir = path.join(worktreePath, '.features');
-      await fs.mkdir(featuresDir, { recursive: true });
-      const indexPath = path.join(featuresDir, 'index.json');
-      try {
-        await fs.access(indexPath);
-      } catch {
-        await fs.writeFile(indexPath, JSON.stringify({ version: 1, features: {} }, null, 2) + '\n');
-      }
-
-      // Remove .disabled sentinel
-      try { await fs.unlink(path.join(featuresDir, '.disabled')); } catch { /* doesn't exist */ }
-
-      // Add SessionEnd hook
-      try {
-        const content = await fs.readFile(settingsPath, 'utf-8');
-        const updated = addKbHook(content, devflowDir);
-        if (updated !== content) {
-          await fs.writeFile(settingsPath, updated, 'utf-8');
-        }
-      } catch { /* settings.json may not exist */ }
-
-      // Update manifest
-      const manifest = await readManifest(devflowDir);
-      if (manifest) {
-        manifest.features.kb = true;
-        manifest.updatedAt = new Date().toISOString();
-        await writeManifest(devflowDir, manifest);
-      }
-
-      p.log.success('Feature knowledge bases enabled');
-      p.outro(`SessionEnd hook installed. Run ${color.cyan('devflow kb create <slug>')} to create a KB.`);
-
-    } else if (options.disable) {
-      p.intro(color.cyan('Disable Feature Knowledge Bases'));
-
-      // Create .disabled sentinel
-      const featuresDir = path.join(worktreePath, '.features');
-      await fs.mkdir(featuresDir, { recursive: true });
-      await fs.writeFile(path.join(featuresDir, '.disabled'), '', 'utf-8');
-
-      // Remove SessionEnd hook
-      try {
-        const content = await fs.readFile(settingsPath, 'utf-8');
-        const updated = removeKbHook(content);
-        if (updated !== content) {
-          await fs.writeFile(settingsPath, updated, 'utf-8');
-        }
-      } catch { /* settings.json may not exist */ }
-
-      // Update manifest
-      const manifest = await readManifest(devflowDir);
-      if (manifest) {
-        manifest.features.kb = false;
-        manifest.updatedAt = new Date().toISOString();
-        await writeManifest(devflowDir, manifest);
-      }
-
-      p.log.success('Feature knowledge bases disabled');
-      p.log.info('Existing KBs preserved. Manual commands (create/refresh) still work.');
-      p.outro('');
-
-    } else {
-      // options.status
-      p.intro(color.cyan('Feature KB Status'));
-
-      // Check hook
-      let hookPresent = false;
-      try {
-        const content = await fs.readFile(settingsPath, 'utf-8');
-        hookPresent = hasKbHook(content);
-      } catch { /* settings.json may not exist */ }
-
-      // Check sentinel
-      let disabled = false;
-      try {
-        await fs.access(path.join(worktreePath, '.features', '.disabled'));
-        disabled = true;
-      } catch { /* not disabled */ }
-
-      // Count KBs
-      const kbs = featureKb.listKBs(worktreePath);
-
-      const enabled = hookPresent && !disabled;
-      p.log.info(`Status: ${enabled ? color.green('enabled') : color.yellow('disabled')}`);
-      p.log.info(`Hook:   ${hookPresent ? color.green('installed') : color.dim('not installed')}`);
-      p.log.info(`KBs:    ${kbs.length}`);
-      if (disabled) {
-        p.log.info(`Sentinel: ${color.yellow('.features/.disabled present')}`);
-      }
-      p.outro('');
-    }
-  });
-
-// ---------------------------------------------------------------------------
-// devflow kb list
-// ---------------------------------------------------------------------------
-
-kbCommand
-  .command('list')
-  .description('List all feature KBs with staleness status')
-  .action(async () => {
-    p.intro(color.cyan('Feature Knowledge Bases'));
-
-    const worktreePath = await getWorktreePath();
-    const kbs = featureKb.listKBs(worktreePath);
-    const staleness = featureKb.checkAllStaleness(worktreePath);
-
-    if (kbs.length === 0) {
-      p.log.info(
-        'No feature KBs found. KBs are created automatically during planning, or manually via ' +
-        color.cyan('devflow kb create <slug>') + '.'
-      );
-      p.outro('');
-      return;
-    }
-
-    p.log.info(`Found ${kbs.length} feature KB${kbs.length === 1 ? '' : 's'} in ${color.dim(worktreePath)}`);
-    console.log('');
-
-    for (const kb of kbs) {
-      const staleInfo = staleness[kb.slug];
-      const isStale = staleInfo?.stale ?? false;
-      const statusBadge = isStale ? color.yellow('[STALE]') : color.green('[current]');
-
-      console.log(`  ${color.bold(kb.name)} ${statusBadge}`);
-      console.log(`    slug:       ${color.dim(kb.slug)}`);
-      console.log(`    updated:    ${color.dim(kb.lastUpdated)}`);
-      console.log(`    dirs:       ${color.dim(kb.directories.join(', '))}`);
-      if (isStale && staleInfo.changedFiles.length > 0) {
-        const shown = staleInfo.changedFiles.slice(0, 3).join(', ');
-        const overflow = staleInfo.changedFiles.length > 3 ? ` +${staleInfo.changedFiles.length - 3} more` : '';
-        console.log(`    changed:    ${color.yellow(shown)}${overflow}`);
-      }
-      console.log('');
-    }
-
-    p.outro(`Run ${color.cyan('devflow kb check')} to see staleness details`);
-  });
-
-// ---------------------------------------------------------------------------
-// devflow kb check
-// ---------------------------------------------------------------------------
-
-kbCommand
-  .command('check')
-  .description('Check all KBs for staleness')
-  .action(async () => {
-    p.intro(color.cyan('KB Staleness Check'));
-
-    const worktreePath = await getWorktreePath();
-    const kbs = featureKb.listKBs(worktreePath);
-    const staleness = featureKb.checkAllStaleness(worktreePath);
-
-    if (kbs.length === 0) {
-      p.log.info('No feature KBs found.');
-      p.outro('');
-      return;
-    }
-
-    let staleCount = 0;
-
-    for (const kb of kbs) {
-      const staleInfo = staleness[kb.slug];
-      const isStale = staleInfo?.stale ?? false;
-      if (isStale) {
-        staleCount++;
-        p.log.warn(`${kb.name} (${kb.slug}) is stale`);
-        for (const f of staleInfo.changedFiles.slice(0, 5)) {
-          console.log(`    ${color.yellow('•')} ${f}`);
-        }
-        if (staleInfo.changedFiles.length > 5) {
-          console.log(`    ${color.yellow('•')} ...and ${staleInfo.changedFiles.length - 5} more`);
-        }
-      } else {
-        p.log.success(`${kb.name} (${kb.slug}) is current`);
-      }
-    }
-
-    if (staleCount > 0) {
-      p.outro(`${staleCount} KB${staleCount === 1 ? '' : 's'} need refresh. Run: ${color.cyan('devflow kb refresh')}`);
-    } else {
-      p.outro('All KBs are current');
-    }
-  });
-
-// ---------------------------------------------------------------------------
-// devflow kb create <slug>
-// ---------------------------------------------------------------------------
-
-kbCommand
-  .command('create <slug>')
-  .description('Create a new KB via claude -p exploration')
-  .action(async (slug: string) => {
-    exitOnInvalidSlug(slug);
-    p.intro(color.cyan(`Create Feature KB: ${slug}`));
-
-    if (!isClaudeCliAvailable()) {
-      p.log.error('claude CLI not found on PATH. Install Claude Code first.');
-      process.exit(1);
-    }
-
-    const worktreePath = await getWorktreePath();
-
-    const name = await p.text({
-      message: 'Feature name (human-readable)',
-      placeholder: 'e.g., CLI Command System',
-      validate: (v) => (v.trim().length < 3 ? 'Name must be at least 3 characters' : undefined),
-    });
-    if (p.isCancel(name)) { p.cancel('Cancelled.'); return; }
-
-    const directoriesRaw = await p.text({
-      message: 'Directories (comma-separated, e.g., src/cli/commands/,src/cli/utils/)',
-      placeholder: 'src/feature/',
-      validate: (v) => (v.trim().length === 0 ? 'Enter at least one directory' : undefined),
-    });
-    if (p.isCancel(directoriesRaw)) { p.cancel('Cancelled.'); return; }
-
-    const directories = (directoriesRaw as string).split(',').map((d) => d.trim()).filter(Boolean);
-    const dirList = directories.map((d) => `"${d}"`).join(', ');
-
-    const s = p.spinner();
-    s.start('Creating KB...');
-
-    const prompt = [
-      `You are the Knowledge agent. Create a feature knowledge base for the following area:`,
-      ``,
-      `FEATURE_SLUG: ${slug}`,
-      `FEATURE_NAME: ${name as string}`,
-      `DIRECTORIES: [${dirList}]`,
-      `WORKTREE_PATH: ${worktreePath}`,
-      ``,
-      `Follow the devflow:feature-kb skill's 4-phase process:`,
-      `1. Scan the directories to identify key files and entry points`,
-      `2. Extract architecture, conventions, patterns, integration points`,
-      `3. Distill into actionable cross-cutting knowledge`,
-      `4. Write .features/${slug}/KNOWLEDGE.md with all required sections`,
-      ``,
-      `After writing KNOWLEDGE.md, write .features/${slug}/.create-result.json with:`,
-      `{`,
-      `  "referencedFiles": [<5-10 key files from the explored directories for staleness tracking>],`,
-      `  "description": "<one-line description starting with 'Use when' for relevance matching>"`,
-      `}`,
-      ``,
-      `Create the directory if needed. Report KB_STATUS when done.`,
-    ].join('\n');
-
-    try {
-      const { sidecar } = await runKbAgent({ worktreePath, slug, prompt, sidecarName: '.create-result.json' });
-
-      featureKb.updateIndex(worktreePath, {
-        slug,
-        name: name as string,
-        directories,
-        referencedFiles: sidecar.referencedFiles ?? [],
-        description: sidecar.description,
-        createdBy: 'devflow-kb',
-      });
-
-      s.stop('KB created successfully');
-      p.log.success(`KB written to .features/${slug}/KNOWLEDGE.md`);
-    } catch (err) {
-      s.stop('KB creation failed');
-      p.log.error(`claude exited with error: ${err instanceof Error ? err.message : String(err)}`);
-      process.exit(1);
-    }
-
-    p.outro(`Run ${color.cyan(`devflow kb list`)} to see all KBs`);
-  });
-
-// ---------------------------------------------------------------------------
-// devflow kb refresh [slug]
-// ---------------------------------------------------------------------------
-
-kbCommand
-  .command('refresh [slug]')
-  .description('Refresh stale KB(s). Omit slug to refresh all stale KBs.')
-  .action(async (slug?: string) => {
-    p.intro(color.cyan(slug ? `Refresh KB: ${slug}` : 'Refresh Stale KBs'));
-
-    if (slug) exitOnInvalidSlug(slug);
-
-    if (!isClaudeCliAvailable()) {
-      p.log.error('claude CLI not found on PATH. Install Claude Code first.');
-      process.exit(1);
-    }
-
-    const worktreePath = await getWorktreePath();
-
-    // Determine which slugs to refresh
-    let slugsToRefresh: string[];
-    let stalenessMap: Record<string, { stale: boolean; changedFiles: string[] }> | undefined;
-    if (slug) {
-      slugsToRefresh = [slug];
-    } else {
-      stalenessMap = featureKb.checkAllStaleness(worktreePath);
-      slugsToRefresh = Object.entries(stalenessMap)
-        .filter(([, info]) => info.stale)
-        .map(([s]) => s);
-    }
-
-    if (slugsToRefresh.length === 0) {
-      p.log.success('No stale KBs found — everything is current.');
-      p.outro('');
-      return;
-    }
-
-    p.log.info(`Refreshing ${slugsToRefresh.length} KB${slugsToRefresh.length === 1 ? '' : 's'}: ${slugsToRefresh.join(', ')}`);
-
-    const kbs = featureKb.listKBs(worktreePath);
-
-    for (const kbSlug of slugsToRefresh) {
-      const s = p.spinner();
-      s.start(`Refreshing ${kbSlug}...`);
-
-      const staleInfo = stalenessMap?.[kbSlug] ?? featureKb.checkStaleness(worktreePath, kbSlug);
-      const kbEntry = kbs.find((k: { slug: string }) => k.slug === kbSlug);
-      const featureName = kbEntry?.name ?? kbSlug;
-      const kbDirectories = kbEntry?.directories ?? [];
-
-      const prompt = [
-        `You are the Knowledge agent refreshing a stale feature knowledge base.`,
-        ``,
-        `FEATURE_SLUG: ${kbSlug}`,
-        `FEATURE_NAME: ${featureName}`,
-        `DIRECTORIES: ${JSON.stringify(kbDirectories)}`,
-        `WORKTREE_PATH: ${worktreePath}`,
-        `CHANGED_FILES: ${JSON.stringify(staleInfo.changedFiles)}`,
-        ``,
-        `Instructions:`,
-        `- Read .features/${kbSlug}/KNOWLEDGE.md to see the existing KB content`,
-        `- Read the CHANGED_FILES to understand what changed`,
-        `- Update the stale sections based on changes`,
-        `- Preserve any manually added content`,
-        `- Do not regenerate from scratch`,
-        `- Write the updated KB to .features/${kbSlug}/KNOWLEDGE.md`,
-        `- Write .features/${kbSlug}/.refresh-result.json with: {"referencedFiles": [<5-10 key files from explored directories for staleness tracking>]}`,
-      ].join('\n');
-
-      try {
-        const { sidecar } = await runKbAgent({ worktreePath, slug: kbSlug, prompt, sidecarName: '.refresh-result.json' });
-
-        featureKb.updateIndex(worktreePath, {
-          slug: kbSlug,
-          name: featureName,
-          directories: kbDirectories,
-          referencedFiles: sidecar.referencedFiles ?? kbEntry?.referencedFiles ?? [],
-          createdBy: 'devflow-kb',
-        });
-
-        s.stop(`${kbSlug} refreshed`);
-      } catch (err) {
-        s.stop(`${kbSlug} refresh failed`);
-        p.log.error(`Error: ${err instanceof Error ? err.message : String(err)}`);
-      }
-    }
-
-    p.outro('Refresh complete');
-  });
-
-// ---------------------------------------------------------------------------
-// devflow kb remove <slug>
-// ---------------------------------------------------------------------------
-
-kbCommand
-  .command('remove <slug>')
-  .description('Remove a KB and its index entry')
-  .action(async (slug: string) => {
-    exitOnInvalidSlug(slug);
-    p.intro(color.cyan(`Remove KB: ${slug}`));
-
-    const confirmed = await p.confirm({
-      message: `Remove KB '${slug}' and its KNOWLEDGE.md? This cannot be undone.`,
-      initialValue: false,
-    });
-    if (p.isCancel(confirmed) || !confirmed) {
-      p.cancel('Removal cancelled.');
-      return;
-    }
-
-    const worktreePath = await getWorktreePath();
-
-    try {
-      featureKb.removeEntry(worktreePath, slug);
-      p.log.success(`KB '${slug}' removed`);
-    } catch (err) {
-      p.log.error(`Failed to remove KB: ${err instanceof Error ? err.message : String(err)}`);
-      process.exit(1);
-    }
-
-    p.outro('Done');
-  });
+export * from './kb/index.js';
